requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
fastapi==0.116.1
gTTS==2.5.4
h11==0.16.0
idna==3.10
lxml==6.0.1
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
PyPDF2==3.0.1
python-docx==1.2.0
python-multipart==0.0.20
reportlab==4.4.3
requests==2.32.5
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
